- replace anonymous functions in classroom component
