+++ app/components/classroom/ClassList.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import {connect} from 'react-redux';
 import {withTranslation} from 'react-i18next';
 import {bindActionCreators} from 'redux';
+
 import {actionCreators} from '../../actions/index';
 import RoomModal from './RoomModal';
-import {sortData, classInputs} from './helpers/formHelpers';
+import {sortData, classInputs} from './formHelpers';
 import css from './styles/room.css';
 
-const ClassList = ({t, listData, actions}) => {
+function ClassList({t, listData, actions}) {
     const cleanedData = sortData(listData);
-
     const listInputs = classInputs(cleanedData, actions.roomModalDisplay);
 
     return (
@@ -19,7 +19,8 @@ const ClassList = ({t, listData, actions}) => {
             <div className='list-group list-group-flush'>{listInputs}</div>
         </div>
     );
-};
+}
+
 const mapDispatchToProps = dispatch => ({
     actions: bindActionCreators(actionCreators, dispatch)
 });
