+++ app/components/classroom/Classes.jsx
@@ -2,17 +2,14 @@ import React from 'react';
 import {connect} from 'react-redux';
 import {withTranslation} from 'react-i18next';
 import {bindActionCreators} from 'redux';
+
 import {actionCreators} from '../../actions/index';
 import ClassList from './ClassList';
 import Subjects from '../subject/Subjects';
-import {
-    addRoomForm,
-    createFormInputs,
-    checkChange
-} from './helpers/formHelpers';
+import {addRoomForm, createFormInputs, checkChange} from './formHelpers';
 import css from './styles/room.css';
 
-const Classes = ({t, classData, actions}) => {
+function Classes({t, classData, actions}) {
     const formInputs = createFormInputs(t, classData);
 
     checkChange(classData, actions);
@@ -30,7 +27,8 @@ const Classes = ({t, classData, actions}) => {
             <Subjects t={t} classData={classData} />
         </div>
     );
-};
+}
+
 const mapStateToProps = state => ({
     classData: state.classData
 });
@@ -39,7 +37,4 @@ const mapDispatchToProps = dispatch => ({
     actions: bindActionCreators(actionCreators, dispatch)
 });
 
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(withTranslation()(Classes));
+export default connect(mapStateToProps, mapDispatchToProps)(withTranslation()(Classes));
