+++ app/components/classroom/RoomModal.jsx
@@ -3,25 +3,20 @@ import _ from 'lodash';
 import {connect} from 'react-redux';
 import {withTranslation} from 'react-i18next';
 import {bindActionCreators} from 'redux';
+
 import {actionCreators} from '../../actions/index';
 import {modalFrame} from '../helpers/editModal';
-import {filterObjectData, createModalInputs} from './helpers/formHelpers';
+import {filterObjectData, createModalInputs} from './formHelpers';
 
-const getCurrentModalData = CurrentModalData => {
+function getCurrentModalData(CurrentModalData) {
     return _.pick(CurrentModalData, ['name', 'teacher', 'substitute']);
-};
+}
 
-const RoomModal = ({t, modalData, classModalData, actions}) => {
+function RoomModal({t, modalData, classModalData, actions}) {
     const {id, showModal, isInvalid} = classModalData;
-
-    const selectedRoom = isInvalid
-        ? getCurrentModalData(classModalData)
-        : filterObjectData(modalData, id);
-
+    const selectedRoom = isInvalid ? getCurrentModalData(classModalData) : filterObjectData(modalData, id);
     const roomInputs = createModalInputs(t, selectedRoom, isInvalid);
-
-    const hiddenInput =
-		<input type='hidden' name='oldName' data-id={selectedRoom.name} />;
+    const hiddenInput = <input type='hidden' name='oldName' data-id={selectedRoom.name} />;
 
     const footerData = {
         dataId:       id,
@@ -43,7 +38,7 @@ const RoomModal = ({t, modalData, classModalData, actions}) => {
             )}
         </div>
     );
-};
+}
 
 const mapStateToProps = state => ({classModalData: state.classModalData});
 
@@ -51,7 +46,4 @@ const mapDispatchToProps = dispatch => ({
     actions: bindActionCreators(actionCreators, dispatch)
 });
 
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(withTranslation()(RoomModal));
+export default connect(mapStateToProps, mapDispatchToProps)(withTranslation()(RoomModal));
