+++ app/components/classroom/reducers/index.js
