+++ app/components/classroom/styles/room.css
@@ -48,7 +48,7 @@
 }
 
 .room_div {
-  white-space: 800px;
+  white-space: normal;
   padding: 0 20px 0 0;
 }
