+++ app/reducers/index.js
@@ -1,7 +1,7 @@
 import {combineReducers} from 'redux';
 import {connectRouter} from 'connected-react-router';
 import {applyAddressData, applyGradeSystem} from '../components/settings/reducers';
-import {applyTabChange, applyClassData, applyClassModal} from '../components/rooms/reducers';
+import {applyTabChange, applyClassData, applyClassModal} from '../components/classroom/reducers';
 import {applyClassList, applySubjectData, applySubjectModal} from '../components/subject/reducers';
 import {applyFilteredExam} from '../components/exam/reducers';
 import {applyStudentData} from '../components/students/reducers';
