+++ app/store/configureStore.dev.js
@@ -4,7 +4,7 @@ import {createHashHistory} from 'history';
 import {routerMiddleware, routerActions} from 'connected-react-router';
 import {createLogger} from 'redux-logger';
 import createRootReducer from '../reducers';
-import * as classroomActions from '../components/rooms/actions';
+import * as classroomActions from '../components/classroom/actions';
 
 const history = createHashHistory();
 const rootReducer = createRootReducer(history);
