+++ package.json
@@ -31,6 +31,38 @@
   "start-main-dev": "cross-env START_HOT=1 NODE_ENV=development electron -r @babel/register ./app/main.dev.js",
   "start-renderer-dev": "cross-env NODE_ENV=development webpack-dev-server --config configs/webpack.config.renderer.dev.babel.js"
  },
+ "dependencies": {
+  "@fortawesome/fontawesome-free": "^5.12.0",
+  "@hot-loader/react-dom": "^16.11.0",
+  "chart.js": "^2.7.3",
+  "connected-react-router": "^6.6.1",
+  "core-js": "^3.4.5",
+  "devtron": "^1.4.0",
+  "electron-debug": "^3.0.1",
+  "electron-log": "^3.0.9",
+  "electron-updater": "^4.2.0",
+  "history": "^4.10.1",
+  "i18next": "^19.1.0",
+  "i18next-electron-language-detector": "^0.0.10",
+  "jspdf": "^1.5.3",
+  "lodash": "^4.17.19",
+  "nedb": "https://github.com/hcuffy/nedb",
+  "react": "^16.12.0",
+  "react-chartjs-2": "^2.7.4",
+  "react-dom": "^16.12.0",
+  "react-hot-loader": "^4.12.18",
+  "react-i18next": "^11.3.1",
+  "react-redux": "^7.1.3",
+  "react-router": "^5.1.2",
+  "react-router-dom": "^5.1.2",
+  "react-table": "^6.8.6",
+  "react-toastify": "^4.3.2",
+  "reactstrap": "^8.2.0",
+  "redux": "^4.0.4",
+  "redux-thunk": "^2.3.0",
+  "source-map-support": "^0.5.16",
+  "thenify": "^3.3.1"
+ },
  "devDependencies": {
   "@babel/core": "^7.7.5",
   "@babel/plugin-proposal-class-properties": "^7.7.4",
@@ -103,38 +135,6 @@
   "webpack-dev-server": "^3.9.0",
   "webpack-merge": "^4.2.2",
   "yarn": "^1.22.0"
- },
- "dependencies": {
-  "@fortawesome/fontawesome-free": "^5.12.0",
-  "@hot-loader/react-dom": "^16.11.0",
-  "chart.js": "^2.7.3",
-  "connected-react-router": "^6.6.1",
-  "core-js": "^3.4.5",
-  "devtron": "^1.4.0",
-  "electron-debug": "^3.0.1",
-  "electron-log": "^3.0.9",
-  "electron-updater": "^4.2.0",
-  "history": "^4.10.1",
-  "i18next": "^19.1.0",
-  "i18next-electron-language-detector": "^0.0.10",
-  "jspdf": "^1.5.3",
-  "lodash": "^4.17.19",
-  "nedb": "https://github.com/hcuffy/nedb",
-  "react": "^16.12.0",
-  "react-chartjs-2": "^2.7.4",
-  "react-dom": "^16.12.0",
-  "react-hot-loader": "^4.12.18",
-  "react-i18next": "^11.3.1",
-  "react-redux": "^7.1.3",
-  "react-router": "^5.1.2",
-  "react-router-dom": "^5.1.2",
-  "react-table": "^6.8.6",
-  "react-toastify": "^4.3.2",
-  "reactstrap": "^8.2.0",
-  "redux": "^4.0.4",
-  "redux-thunk": "^2.3.0",
-  "source-map-support": "^0.5.16",
-  "thenify": "^3.3.1"
  },
  "build": {
   "productName": "Marks",
